+++ spec/spec_helper.rb
@@ -2,3 +2,4 @@ require 'simplecov'
 SimpleCov.start
 
 require "pry"
+RSpec::Mocks.configuration.allow_message_expectations_on_nil = true